JDK-8200224
Moved JFXPanelTest to the system tests. The new test is now part of the already existing test class.
